+++ package.json
@@ -31,6 +31,7 @@
     "react": "^16.4",
     "react-dom": "^16.4",
     "react-items-carousel": "^2.8.0",
+    "react-share": "^4.4.0",
     "react-simply-carousel": "^5.1.3"
   }
 }
